wip(page-elements): use page elements if possible
